Update jacobi_iteration_method.py
Fixed bug: init_val: list[int]  to  list[float].
Since the numbers are fractional: init_val = [0.5, -0.5, -0.5].

+++ arithmetic_analysis/jacobi_iteration_method.py
@@ -12,7 +12,7 @@ from numpy.typing import NDArray
 def jacobi_iteration_method(
     coefficient_matrix: NDArray[float64],
     constant_matrix: NDArray[float64],
-    init_val: list[int],
+    init_val: list[float],
     iterations: int,
 ) -> list[float]:
     """
